fix: move template.ts to turbo/scripts/e2b for module resolution
Move template.ts from e2b/ to turbo/scripts/e2b/ so it can access
node_modules from the turbo workspace. This fixes the 'Cannot find
module e2b' error.

Changes:
- Move e2b/template.ts to turbo/scripts/e2b/template.ts
- Update imports in build scripts to use './template'
- Add turbo/scripts/e2b/** to workflow path triggers
