+++ .github/workflows/e2b-template.yml
@@ -6,10 +6,12 @@ on:
       - main
     paths:
       - "e2b/**"
+      - "turbo/scripts/e2b/**"
       - ".github/workflows/e2b-template.yml"
   pull_request:
     paths:
       - "e2b/**"
+      - "turbo/scripts/e2b/**"
       - ".github/workflows/e2b-template.yml"
 
 jobs:
